Hoist/subset configs
